client: Give each participant an integer identifier.
This change gives each participant an integer identifier which is
also provided to them via a sound when they join. This allows them
to watch themselves without exposing personal information such as
their phone number.

Right now there is a problem with sound playback to participants
where it may be out of order and can't be stopped if need be. This
will be fixed.